Refactor createConfig into an independent file for now
